+++ iexfinance/__init__.py
@@ -3,13 +3,16 @@ from iexfinance.base import _IEXBase
 from iexfinance.market import TOPS, Last, DEEP, Book
 from iexfinance.stats import (IntradayReader, RecentReader, RecordsReader,
                               DailySummaryReader, MonthlySummaryReader)
-from iexfinance.ref import (CorporateActions, Dividends, NextDay,
-                            ListedSymbolDir, Symbols, IEXSymbols)
+from iexfinance.refdata import (CorporateActions, Dividends, NextDay,
+                                ListedSymbolDir)
 from iexfinance.utils import _sanitize_dates
 from iexfinance.utils.exceptions import IEXQueryError
 
 __author__ = 'Addison Lynch'
 __version__ = '0.3.5'
+
+WNG_MSG = "%s is moved to iexfinance.%s. This funciton will be "\
+          "deprecated in 0.4.1."
 
 # Data provided for free by IEX
 # Data is furnished in compliance with the guidelines promulgated in the IEX
@@ -19,39 +22,59 @@ __version__ = '0.3.5'
 
 
 def get_market_gainers(*args, **kwargs):
+    """
+    MOVED to iexfinance.stocks.get_market_gainers
+    """
+    import warnings
+    warnings.warn(WNG_MSG, ("get_market_gainers", "stocks.get_market_gainers"))
     return stocks.get_market_gainers(*args, **kwargs)
 
 
 def get_market_losers(*args, **kwargs):
+    """
+    MOVED to iexfinance.stocks.get_market_losers
+    """
+    import warnings
+    warnings.warn(WNG_MSG, ("get_market_losers", "stocks.get_market_losers"))
     return stocks.get_market_losers(*args, **kwargs)
 
 
 def get_market_most_active(*args, **kwargs):
+    """
+    MOVED to iexfinance.stocks.get_market_most_active
+    """
+    import warnings
+    warnings.warn(WNG_MSG, ("get_market_most_active",
+                            "stocks.get_market_most_active"))
     return stocks.get_market_most_active(*args, **kwargs)
 
 
 def get_market_iex_volume(*args, **kwargs):
+    """
+    MOVED to iexfinance.stocks.get_market_iex_volume
+    """
+    import warnings
+    warnings.warn(WNG_MSG, ("get_market_iex_volume",
+                            "stocks.get_market_iex_volume"))
     return stocks.get_market_iex_volume(*args, **kwargs)
 
 
 def get_market_iex_percent(*args, **kwargs):
+    """
+    MOVED to iexfinance.stocks.get_market_iex_percent
+    """
+    import warnings
+    warnings.warn(WNG_MSG, ("get_market_iex_percent",
+                            "stocks.get_market_iex_percent"))
     return stocks.get_market_iex_percent(*args, **kwargs)
 
 
 def get_available_symbols(**kwargs):
     """
-    Top-level function to obtain IEX available symbols
-
-    Parameters
-    ----------
-    kwargs:
-        Additional Request Parameters (see base class)
-
-    Returns
-    -------
-    data: list
-        List of dictionary reference items for each symbol
+    MOVED to iexfinance.refdata.get_symbols
     """
+    import warnings
+    warnings.warn(WNG_MSG % ("get_available_symbols", "refdata.get_symbols"))
     _ALL_SYMBOLS_URL = "https://api.iextrading.com/1.0/ref-data/symbols"
     handler = _IEXBase(**kwargs)
     response = handler._execute_iex_query(_ALL_SYMBOLS_URL)
@@ -63,134 +86,64 @@ def get_available_symbols(**kwargs):
 
 def get_iex_corporate_actions(start=None, **kwargs):
     """
-    Top-level function to retrieve IEX Corporate Actions from the ref-data
-    endpoints
-
-    Parameters
-    ----------
-    start: datetime.datetime, default None, optional
-        A month to use for retrieval (a datetime object)
-    kwargs: Additional Request Parameters (see base class)
+    MOVED to iexfinance.refdata.get_iex_corporate_actions
     """
+    import warnings
+    warnings.warn(WNG_MSG % ("get_iex_corporate_actions",
+                             "refdata.get_iex_corporate_actions"))
     return CorporateActions(start=start, **kwargs).fetch()
 
 
 def get_iex_dividends(start=None, **kwargs):
     """
-    Top-level function to retrieve IEX Dividends from the ref-data
-    endpoints
-
-    Parameters
-    ----------
-    start: datetime.datetime, default None, optional
-        A month to use for retrieval (a datetime object)
-    kwargs: Additional Request Parameters (see base class)
+    MOVED to iexfinance.refdata.get_iex_dividends
     """
+    import warnings
+    warnings.warn(WNG_MSG % ("get_iex_dividends", "refdata.get_iex_dividends"))
     return Dividends(start=start, **kwargs).fetch()
 
 
 def get_iex_next_day_ex_date(start=None, **kwargs):
     """
-    Top-level function to retrieve IEX Next Day Ex Date from the ref-data
-    endpoints
-
-    Parameters
-    ----------
-    start: datetime.datetime, default None, optional
-        A month to use for retrieval (a datetime object)
-    kwargs: Additional Request Parameters (see base class)
+    MOVED to iexfinance.refdata.get_iex_next_day_ex_date
     """
+    import warnings
+    warnings.warn(WNG_MSG % ("get_iex_next_day_ex_date",
+                             "refdata.get_iex_next_day_ex_date"))
     return NextDay(start=start, **kwargs).fetch()
 
 
 def get_iex_listed_symbol_dir(start=None, **kwargs):
     """
-    Top-level function to retrieve IEX Listed Symbol Directory from the
-    ref-data endpoints
-
-    Parameters
-    ----------
-    start: datetime.datetime, default None, optional
-        A month to use for retrieval (a datetime object)
-    kwargs: Additional Request Parameters (see base class)
+    MOVED to iexfinance.refdata.get_symbols
     """
+    import warnings
+    warnings.warn(WNG_MSG % ("get_iex_listed_symbol_dir",
+                             "refdata.get_iex_listed_symbol_dir"))
     return ListedSymbolDir(start=start, **kwargs)
-
-
-# IEX Cloud reference endpoints
-
-def get_symbols(**kwargs):
-    """
-    Top-level function to retrieve array of all symbols that IEX Cloud supports
-    for API calls
-
-    Reference: https://iexcloud.io/docs/api/#symbols
-
-    .. warning:: This endpoint is only available using IEX Cloud. See
-                 :ref:`Migrating` for more information.
-
-    Data Weighting: ``100`` per call
-    """
-    return Symbols(**kwargs).fetch()
-
-
-def get_iex_symbols(**kwargs):
-    """
-    Top-level function to retrieve array of all symbols the Investor's Exchange
-    supports for trading
-
-    Reference: https://iexcloud.io/docs/api/#iex-symbols
-
-    .. warning:: This endpoint is only available using IEX Cloud. See
-                 :ref:`Migrating` for more information.
-
-    Data Weighting: ``Free``
-    """
-    return IEXSymbols(**kwargs).fetch()
 
 
 def get_market_tops(symbols=None, **kwargs):
     """
-    Top-level function to obtain TOPS data for a symbol or list of symbols
-
-    Parameters
-    ----------
-    symbols: str or list, default None, optional
-        A symbol or list of symbols
-    kwargs:
-        Additional Request Parameters (see base class)
+    MOVED to iexfinance.iexdata.get_tops
     """
+    import warnings
+    warnings.warn(WNG_MSG % ("get_market_tops", "iexdata.get_tops"))
     return TOPS(symbols, **kwargs).fetch()
 
 
 def get_market_last(symbols=None, **kwargs):
     """
-    Top-level function to obtain Last data for a symbol or list of symbols
-
-    Parameters
-    ----------
-    symbols: str or list, default None, optional
-        A symbol or list of symbols
-    kwargs:
-        Additional Request Parameters (see base class)
+    MOVED to iexfinance.iexdata.get_last
     """
+    import warnings
+    warnings.warn(WNG_MSG % ("get_market_last", "iexdata.get_last"))
     return Last(symbols, **kwargs).fetch()
 
 
 def get_market_deep(symbols=None, **kwargs):
     """
-    Top-level function to obtain DEEP data for a symbol or list of symbols
-
-    Parameters
-    ----------
-    symbols: str or list, default None
-        A symbol or list of symbols
-    kwargs:
-        Additional Request Parameters (see base class)
-
-    Notes
-    -----
-    Pandas not supported as an output format for the DEEP endpoint.
+    MOVED to iexfinance.iexdata.get_deep
     """
     return DEEP(symbols, **kwargs).fetch()
