+++ iexfinance/refdata/base.py
@@ -1,7 +1,7 @@
 import datetime
 
-from .base import _IEXBase
-from iexfinance.utils import cloud_endpoint
+from iexfinance.base import _IEXBase
+from iexfinance.utils import cloud_endpoint, legacy_endpoint
 
 # Data provided for free by IEX
 # See https://iextrading.com/api-exhibit-a/ for additional information
@@ -31,12 +31,20 @@ class CorporateActions(ReferenceReader):
     def endpoint(self):
         return 'corporate-actions'
 
+    @legacy_endpoint
+    def fetch(self):
+        return super(CorporateActions, self).fetch()
+
 
 class Dividends(ReferenceReader):
 
     @property
     def endpoint(self):
         return 'dividends'
+
+    @legacy_endpoint
+    def fetch(self):
+        return super(Dividends, self).fetch()
 
 
 class NextDay(ReferenceReader):
@@ -45,12 +53,20 @@ class NextDay(ReferenceReader):
     def endpoint(self):
         return 'next-day-ex-date'
 
+    @legacy_endpoint
+    def fetch(self):
+        return super(NextDay, self).fetch()
+
 
 class ListedSymbolDir(ReferenceReader):
 
     @property
     def endpoint(self):
         return 'symbol-directory'
+
+    @legacy_endpoint
+    def fetch(self):
+        return super(CorporateActions, self).fetch()
 
 
 class CloudRef(_IEXBase):
